CME: _EKA_CHECK_BOOK_INTEGRITY ifdef for self check

+++ apisrc/EkaFhCmeGr.h
@@ -70,7 +70,8 @@ public:
   using SizeT       = int32_t;
 
   using SequenceT   = uint32_t;
-  using PriceLevetT = uint8_t;
+  //  using PriceLevetT = uint8_t;
+  using PriceLevetT = ssize_t;
 
   using FhSecurity  = EkaFhCmeSecurity  <
     PriceLevetT,
